requirements.txt
asgiref==3.6.0
Django==4.1.7
django-ninja==0.20.0
pip==23.0
pydantic==1.10.5
PyMySQL==1.0.2
setuptools==65.5.0
sqlparse==0.4.3
typing_extensions==4.5.0
restframework